+++ exasol/analytics/query_handler/graph/stage/stage_graph.py
@@ -1,4 +1,4 @@
-from exasol.analytics.query_handler.execution_graph import ExecutionGraph
+from exasol.analytics.query_handler.graph.execution_graph import ExecutionGraph
 from exasol.analytics.query_handler.graph.stage.stage import Stage
 
 StageGraph = ExecutionGraph[Stage]
