+++ tests/unit_tests/sql_stage_graph/stage_graph_execution_query_handler/test_state_reference_counting.py
@@ -7,7 +7,7 @@ from exasol.analytics.query_handler.result import Finish, Continue
 
 from exasol.analytics.query_handler.graph.stage.sql.sql_stage_graph import SQLStageGraph
 from exasol.analytics.query_handler.graph.stage.sql.execution.query_handler_state import ResultHandlerReturnValue
-from tests.unit_tests.sql_stage_graph.mock_cast import mock_cast
+from tests.mock_cast import mock_cast
 from tests.unit_tests.sql_stage_graph.stage_graph_execution_query_handler.state_test_setup import TestSetup, \
     create_mocks_for_stage, create_execution_query_handler_state_setup
