+++ tests/unit_tests/sql_stage_graph/test_dataset.py
@@ -11,9 +11,9 @@ from exasol.analytics.schema import (
     ColumnType,
 )
 
-from exasol.analytics.query_handler.graph.stage.sql.execution.data_partition import DataPartition
-from exasol.analytics.query_handler.graph.stage.sql.execution.dataset import Dataset
-from exasol.analytics.query_handler.graph.stage.sql.execution.dependency import Dependencies
+from exasol.analytics.query_handler.graph.stage.sql.data_partition import DataPartition
+from exasol.analytics.query_handler.graph.stage.sql.dataset import Dataset
+from exasol.analytics.query_handler.graph.stage.sql.dependency import Dependencies
 
 
 class TestEnum(Enum):
