+++ tests/unit_tests/sql_stage_graph/test_sql_stage_train_query_handler_input.py
@@ -4,7 +4,7 @@ from unittest.mock import create_autospec, MagicMock
 import pytest
 from exasol_bucketfs_utils_python.abstract_bucketfs_location import AbstractBucketFSLocation
 
-from exasol.analytics.query_handler.graph.stage.sql.execution.input_output import SQLStageInputOutput
+from exasol.analytics.query_handler.graph.stage.sql.input_output import SQLStageInputOutput
 from exasol.analytics.query_handler.graph.stage.sql.sql_stage_train_query_handler import SQLStageTrainQueryHandlerInput
